Add test coverage for Travis CI.
Currently, Travis will only attempt to build the project through Makefiles, and
thereby reveal problems related to the compilation process. It will not run the
examples in the tests directory due to an error that prevents the examples to be
properly executed on a Linux box.

This commit adds the tests for the examples, along with a configuration file for
Travis CI and a setup script.