Move slothc to the shared directory.
This also introduces an ext variable for defining the file extensions for the
out files compiled by slothc. Both compilers include this variable in
compiler/codegen.ml, and it is set to "slg" and "sls" for the G machine and
Sestoft machine, respectively. Additionally, the Makefiles have been changed to
link to slothc now located in the shared directory.

The compiler.ml has been changed, so the standard functions (add, if, eq, etc.)
are included when calling the compProg function in order to conform with the
slothc module.

Finally, the compile_examples script have been removed in favour of a new task
in the Makefile in the root of the project called examples. It compiles the
examples for both machines.