Use `getopt` to parse command line arguments.
As we continuously add more options to the stack machine, it becomes cumbersome
to handle each one, and we increase the risk of introducing errors related to
array indices. This commit changes the way we parse command line arguments by
using `getopt` instead.

It also changes the `--verbose` flag to `-v`, and the `--manualstep` flag to
simply `-m`. The latter now sets both the verbose flag, and the manual step
flag.

See more about `getopt` here:
http://www.gnu.org/software/libc/manual/html_node/Example-of-Getopt.html